Update removeUser function to check if the email matches any users

+++ src/hooks/useRemoveUser.tsx
@@ -1,17 +1,18 @@
 import { deleteDoc, collection, query, where, getDocs} from "firebase/firestore";
 import { database } from "../services/firestoreService";
 
-const useRemoveUser = async (email:string) => {     
+const removeUser = async (email:string) => {     
     try {
         if(email){
             const collectionRef = collection(database, "test_users1");
             const q= query(collectionRef, where("email", "==", email));
             const querySnapshot = await getDocs(q);
-
-            querySnapshot.forEach(async user => {
-                await deleteDoc(user.ref)
-
-            })
+            if(!querySnapshot.empty){
+                return querySnapshot.forEach(async user => {
+                    await deleteDoc(user.ref)
+                })
+            }
+            return "Invalid user"    
         }
     }catch(error){
         console.error("Error deleting documents:", error);
@@ -19,4 +20,4 @@ const useRemoveUser = async (email:string) => {
     }
 }
 
-export default useRemoveUser;
+export default removeUser;
